requirements.txt
elasticsearch>=8.11.0
pyyaml>=6.0
urwid>=2.1.2
requests>=2.28.0
